reformatted multitask testing and functionality
